requirements.txt
# requirements.txt
streamlit>=1.28.0
rdflib>=7.0.0
pyvis>=0.3.2
plotly>=5.17.0
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pyyaml>=6.0
streamlit-aggrid>=0.3.4